smolagent_tutorial/requirements.txt
langchain 
google-search-results
smolagents
accelerate
python-dotenv